@@ -1,38 +1,16 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using ModelContextProtocol.Server;
-using System.ComponentModel;
 
-// Create a generic host builder for
-// dependency injection, logging, and configuration.
 var builder = Host.CreateApplicationBuilder(args);
 
-// Configure logging for better integration with MCP clients.
-builder.Logging.AddConsole(consoleLogOptions =>
-{
-consoleLogOptions.LogToStandardErrorThreshold = LogLevel.Trace;
-});
+// Configure all logs to go to stderr (stdout is used for the MCP protocol messages).
+builder.Logging.AddConsole(o => o.LogToStandardErrorThreshold = LogLevel.Trace);
 
-// Register the MCP server and configure it to use stdio transport.
-// Scan the assembly for tool definitions.
+// Add the MCP services: the transport to use (stdio) and the tools to register.
 builder.Services
     .AddMcpServer()
     .WithStdioServerTransport()
-    .WithToolsFromAssembly();
+    .WithTools<RandomNumberTools>();
 
-// Build and run the host. This starts the MCP server.
 await builder.Build().RunAsync();
-
-// Define a static class to hold MCP tools.
-[McpServerToolType]
-public static class EchoTool
-{
-    // Expose a tool that echoes the input message back to the client.
-    [McpServerTool, Description("Echoes the message back to the client.")]
-    public static string Echo(string message) => $"Hello from C#: {message}";
-
-    // Expose a tool that returns the input message in reverse.
-    [McpServerTool, Description("Echoes in reverse the message sent by the client.")]
-    public static string ReverseEcho(string message) => new string(message.Reverse().ToArray());
-}
