@@ -1,7 +1,7 @@
-using Microsoft.Extensions.AI;
+﻿using Microsoft.Extensions.AI;
+using OllamaSharp;
 
-IChatClient client = new SampleChatClient(
-    new Uri("http://coolsite.ai"), "target-ai-model");
+IChatClient client = new OllamaApiClient(new Uri("http://localhost:11434"), "llama3.1");
 
 // <Snippet1>
 List<ChatMessage> history = [];
