@@ -5,14 +5,14 @@ SampleCustomer c = new SampleCustomer("customer one", new DateTime(2010, 5, 31))
     Reminders =
     {
         { new DateTime(2010, 08, 12), "childs's birthday" },
-        { new DateTime(1012, 11, 15), "anniversary" }
+        { new DateTime(2012, 11, 15), "anniversary" }
     }
 };
 
 SampleOrder o = new SampleOrder(new DateTime(2012, 6, 1), 5m);
 c.AddOrder(o);
 
-o = new SampleOrder(new DateTime(2103, 7, 4), 25m);
+o = new SampleOrder(new DateTime(2013, 7, 4), 25m);
 c.AddOrder(o);
 
 Console.WriteLine($"Data about {c.Name}");
