@@ -6,7 +6,7 @@ namespace AsyncBreakfast
     class Program
     {
         // These classes are intentionally empty for the purpose of this example. They are simply marker classes for the purpose of demonstration, contain no properties, and serve no other purpose.
-        internal class Bacon { }
+        internal class HashBrown { }
         internal class Coffee { }
         internal class Egg { }
         internal class Juice { }
@@ -21,8 +21,8 @@ namespace AsyncBreakfast
             Egg eggs = await FryEggsAsync(2);
             Console.WriteLine("eggs are ready");
 
-            Bacon bacon = await FryBaconAsync(3);
-            Console.WriteLine("bacon is ready");
+            HashBrown hashBrown = await FryHashBrownsAsync(3);
+            Console.WriteLine("hash browns are ready");
 
             Toast toast = await ToastBreadAsync(2);
             ApplyButter(toast);
@@ -60,20 +60,20 @@ namespace AsyncBreakfast
             return new Toast();
         }
 
-        private static async Task<Bacon> FryBaconAsync(int slices)
+        private static async Task<HashBrown> FryHashBrownsAsync(int patties)
         {
-            Console.WriteLine($"putting {slices} slices of bacon in the pan");
-            Console.WriteLine("cooking first side of bacon...");
+            Console.WriteLine($"putting {patties} hash brown patties in the pan");
+            Console.WriteLine("cooking first side of hash browns...");
             await Task.Delay(3000);
-            for (int slice = 0; slice < slices; slice++)
+            for (int patty = 0; patty < patties; patty++)
             {
-                Console.WriteLine("flipping a slice of bacon");
+                Console.WriteLine("flipping a hash brown patty");
             }
-            Console.WriteLine("cooking the second side of bacon...");
+            Console.WriteLine("cooking the second side of hash browns...");
             await Task.Delay(3000);
-            Console.WriteLine("Put bacon on plate");
+            Console.WriteLine("Put hash browns on plate");
 
-            return new Bacon();
+            return new HashBrown();
         }
 
         private static async Task<Egg> FryEggsAsync(int howMany)
