@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 namespace AsyncBreakfast
 {
     // These classes are intentionally empty for the purpose of this example. They are simply marker classes for the purpose of demonstration, contain no properties, and serve no other purpose.
-    internal class Bacon { }
+    internal class HashBrown { }
     internal class Coffee { }
     internal class Egg { }
     internal class Juice { }
@@ -20,8 +20,8 @@ namespace AsyncBreakfast
             Egg eggs = FryEggs(2);
             Console.WriteLine("eggs are ready");
 
-            Bacon bacon = FryBacon(3);
-            Console.WriteLine("bacon is ready");
+            HashBrown hashBrown = FryHashBrowns(3);
+            Console.WriteLine("hash browns are ready");
 
             Toast toast = ToastBread(2);
             ApplyButter(toast);
@@ -58,20 +58,20 @@ namespace AsyncBreakfast
             return new Toast();
         }
 
-        private static Bacon FryBacon(int slices)
+        private static HashBrown FryHashBrowns(int patties)
         {
-            Console.WriteLine($"putting {slices} slices of bacon in the pan");
-            Console.WriteLine("cooking first side of bacon...");
+            Console.WriteLine($"putting {patties} hash brown patties in the pan");
+            Console.WriteLine("cooking first side of hash browns...");
             Task.Delay(3000).Wait();
-            for (int slice = 0; slice < slices; slice++)
+            for (int patty = 0; patty < patties; patty++)
             {
-                Console.WriteLine("flipping a slice of bacon");
+                Console.WriteLine("flipping a hash brown patty");
             }
-            Console.WriteLine("cooking the second side of bacon...");
+            Console.WriteLine("cooking the second side of hash browns...");
             Task.Delay(3000).Wait();
-            Console.WriteLine("Put bacon on plate");
+            Console.WriteLine("Put hash browns on plate");
 
-            return new Bacon();
+            return new HashBrown();
         }
 
         private static Egg FryEggs(int howMany)
