@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Runtime.InteropServices;
 using Microsoft.Data.Sqlite;
 using SQLitePCL;
 
@@ -13,10 +12,10 @@ namespace SystemLibrarySample
             readonly IntPtr _library;
 
             public NativeLibraryAdapter(string name)
-                => _library = NativeLibrary.Load(name);
+                => _library = SQLitePCL.NativeLibrary.Load(name);
 
             public IntPtr GetFunctionPointer(string name)
-                => NativeLibrary.TryGetExport(_library, name, out var address)
+                => SQLitePCL.NativeLibrary.TryGetExport(_library, name, out var address)
                     ? address
                     : IntPtr.Zero;
         }
